Add DAP tag.
This is the last portion of the Html component we need, so we can also remove
that file. We won't be including newrelic per 18F's recommended
configurations.
