--- ui/components/dap.js
@@ -1,11 +1,0 @@
-import React from 'react';
-
-
-export default () => (
-  <script
-    async
-    type="text/javascript"
-    id="_fed_an_ua_tag"
-    src="https://dap.digitalgov.gov/Universal­Federated­Analytics­Min.js?agency=EOP&subagency=OMB"
-  />
-);
